new functions: normalize and unormalize center and reduce version
There is a need to create these functions because in `utils_training.py`
in `testmodel()` it performs two tasks. The first task is to test the
model and retrieve the predictions, while the second task involves
unnormalizing the data. Combining these two tasks into a single function
is not recommended in terms of good practice.

Signed-off-by: Guillaume COGONI <cogoni.guillaume@gmail.com>